requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
requests
python-dotenv